requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
markdown
